refactor: move exception response utilities
Add exception response building utilities into a public namespace, `litestar.exceptions.responses`.

These utilities are generally more useful than just within the scope of the exception handler middleware, as evidenced by their use in `litestar-fullstack`.
