+++ docs/reference/exceptions.rst
@@ -3,3 +3,6 @@ exceptions
 
 .. automodule:: litestar.exceptions
     :members:
+
+.. automodule:: litestar.exceptions.responses
+    :members:
