+++ litestar/exceptions/responses/templates/frame.html
