+++ litestar/exceptions/responses/templates/scripts.js
