+++ litestar/middleware/_internal/exceptions/middleware.py
@@ -1,19 +1,17 @@
 from __future__ import annotations
 
 import pdb  # noqa: T100
-from dataclasses import asdict, dataclass, field
 from inspect import getmro
 from sys import exc_info
 from traceback import format_exception
 from typing import TYPE_CHECKING, Any, Type, cast
 
-from litestar.enums import MediaType, ScopeType
+from litestar.enums import ScopeType
 from litestar.exceptions import HTTPException, LitestarException, WebSocketException
-from litestar.middleware._internal.exceptions._debug_response import (
-    _get_type_encoders_for_request,
+from litestar.exceptions.responses import create_exception_response
+from litestar.exceptions.responses._debug_response import (
     create_debug_response,
 )
-from litestar.serialization import encode_json, get_serializer
 from litestar.status_codes import HTTP_500_INTERNAL_SERVER_ERROR
 from litestar.utils.deprecation import warn_deprecation
 from litestar.utils.empty import value_or_raise
@@ -37,7 +35,7 @@ if TYPE_CHECKING:
     )
     from litestar.types.asgi_types import WebSocketCloseEvent
 
-__all__ = ("ExceptionHandlerMiddleware", "ExceptionResponseContent", "create_exception_response")
+__all__ = ("ExceptionHandlerMiddleware",)
 
 
 def get_exception_handler(exception_handlers: ExceptionHandlersMap, exc: Exception) -> ExceptionHandler | None:
@@ -75,44 +73,6 @@ def get_exception_handler(exception_handlers: ExceptionHandlersMap, exc: Excepti
     )
 
 
-@dataclass
-class ExceptionResponseContent:
-    """Represent the contents of an exception-response."""
-
-    status_code: int
-    """Exception status code."""
-    detail: str
-    """Exception details or message."""
-    media_type: MediaType | str
-    """Media type of the response."""
-    headers: dict[str, str] | None = field(default=None)
-    """Headers to attach to the response."""
-    extra: dict[str, Any] | list[Any] | None = field(default=None)
-    """An extra mapping to attach to the exception."""
-
-    def to_response(self, request: Request | None = None) -> Response:
-        """Create a response from the model attributes.
-
-        Returns:
-            A response instance.
-        """
-        from litestar.response import Response
-
-        content: Any = {k: v for k, v in asdict(self).items() if k not in ("headers", "media_type") and v is not None}
-        type_encoders = _get_type_encoders_for_request(request) if request is not None else None
-
-        if self.media_type != MediaType.JSON:
-            content = encode_json(content, get_serializer(type_encoders))
-
-        return Response(
-            content=content,
-            headers=self.headers,
-            status_code=self.status_code,
-            media_type=self.media_type,
-            type_encoders=type_encoders,
-        )
-
-
 def _starlette_exception_handler(request: Request[Any, Any, Any], exc: StarletteHTTPException) -> Response:
     return create_exception_response(
         request=request,
@@ -122,54 +82,6 @@ def _starlette_exception_handler(request: Request[Any, Any, Any], exc: Starlette
             headers=exc.headers,
         ),
     )
-
-
-def create_exception_response(request: Request[Any, Any, Any], exc: Exception) -> Response:
-    """Construct a response from an exception.
-
-    Notes:
-        - For instances of :class:`HTTPException <litestar.exceptions.HTTPException>` or other exception classes that have a
-          ``status_code`` attribute (e.g. Starlette exceptions), the status code is drawn from the exception, otherwise
-          response status is ``HTTP_500_INTERNAL_SERVER_ERROR``.
-
-    Args:
-        request: The request that triggered the exception.
-        exc: An exception.
-
-    Returns:
-        Response: HTTP response constructed from exception details.
-    """
-    headers: dict[str, Any] | None
-    extra: dict[str, Any] | list | None
-
-    if isinstance(exc, HTTPException):
-        status_code = exc.status_code
-        headers = exc.headers
-        extra = exc.extra
-    else:
-        status_code = HTTP_500_INTERNAL_SERVER_ERROR
-        headers = None
-        extra = None
-
-    detail = (
-        exc.detail
-        if isinstance(exc, LitestarException) and status_code != HTTP_500_INTERNAL_SERVER_ERROR
-        else "Internal Server Error"
-    )
-
-    try:
-        media_type = request.route_handler.media_type
-    except (KeyError, AttributeError):
-        media_type = MediaType.JSON
-
-    content = ExceptionResponseContent(
-        status_code=status_code,
-        detail=detail,
-        headers=headers,
-        extra=extra,
-        media_type=media_type,
-    )
-    return content.to_response(request=request)
 
 
 class ExceptionHandlerMiddleware:
